backend/app/requirements.txt
fastapi
uvicorn
httpx
smartapi-python
python-jose
python-dotenv
supabase
